change layout to be more clean
