+++ components/Header/styles.module.scss
@@ -1,0 +1,25 @@
+.backgroundHeader {
+  margin: 2rem 0 4rem 0;
+  padding: 0 2rem;
+
+  h1 {
+    margin: 0;
+    padding: 1rem 0 2rem 0;
+    text-align: center;
+    font-size: 2.2rem;
+    line-height: 1.2;
+    font-weight: 800;
+    letter-spacing: -0.05rem;
+  }
+
+  ul {
+    display: flex;
+    flex-direction: row;
+    list-style-type: none;
+    justify-content: center;
+
+    li {
+      width: 20%;
+    }
+  }
+}
