--- components/HeaderHome/styles.module.scss
@@ -1,42 +1,0 @@
-.backgroundHeader {
-  background: url("../../public/gifs/matrix-animation-background.gif") #fff;
-  background-repeat: no-repeat;
-  background-size: 100% 12rem;
-
-  width: 100%;
-  min-width: 30rem;
-  padding-top: 6rem;
-  border-radius: 8px 8px 0 0;
-
-  h1 {
-    margin-bottom: 0;
-    font-size: 2.2rem;
-    line-height: 1.2;
-    font-weight: 800;
-    letter-spacing: -0.05rem;
-  }
-
-  img {
-    border-radius: 9999px;
-    box-shadow: 0px 0px 2px 1px rgba(0, 0, 0, 0.2);
-    transition: box-shadow 0.4s;
-
-    &:hover {
-      cursor: pointer;
-      box-shadow: 4px 4px 2px 1px rgba(0, 0, 0, 0.2);
-      transition: box-shadow 0.4s;
-    }
-  }
-
-  @media (max-width: 34rem) {
-    margin: 0;
-    border-radius: 0;
-  }
-}
-
-.headerHomeImage {
-  width: 11rem;
-  height: 11rem;
-  margin: 0 auto;
-  border: 4px solid rgba(255, 255, 255, 0.5);
-}
