--- components/HeaderPost/styles.module.scss
@@ -1,58 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: row-reverse;
-  align-items: baseline;
-  justify-content: space-between;
-
-  width: 100%;
-  padding: 2rem 5rem 0;
-  margin: 0;
-
-  border-radius: 30px 30px 0 0;
-  background-color: #fff;
-
-  img {
-    border-radius: 9999px;
-    width: 7rem;
-    height: 7rem;
-    margin: 0 auto;
-  }
-
-  .contactLink {
-    transition: transform 4s;
-  }
-
-  .contactLink:hover {
-    transition: transform 4s;
-    transform: rotateZ(360deg);
-  }
-
-  .backToHome {
-    margin: 0;
-    height: 100%;
-
-    a {
-      color: inherit;
-    }
-
-    .moveToLeft {
-      transition: transform 1s;
-    }
-
-    &:hover {
-      .moveToLeft {
-        transform: translate(-1.2rem, 0);
-      }
-    }
-  }
-
-  @media (max-width: 34rem) {
-    padding: 1rem 2rem;
-    border-radius: 0;
-
-    img {
-      width: 8rem;
-      height: 8rem;
-    }
-  }
-}
