+++ components/Layout/styles.module.scss
@@ -5,20 +5,11 @@
   width: 100%;
 
   margin: 1rem auto 0;
-  padding: 0;
-}
-
-.header {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-align: center;
-  margin: 0;
+  padding: 1rem 2rem;
 }
 
 .mainContainer {
-  background: #fff;
-  padding: 2rem 4rem;
+  padding-top: 0;
 }
 
 @media (max-width: 34rem) {
@@ -29,7 +20,6 @@
   }
 
   .mainContainer {
-    padding: 1rem 2rem;
     font-size: 20px;
   }
 }
@@ -41,6 +31,6 @@
 
   width: 100%;
   min-width: 30rem;
-  padding-top: 6rem;
-  margin-bottom: 1rem;
+  padding: 6rem 0 0 0;
+  margin: 4rem 0 1rem 0;
 }
