+++ pages/posts/styles.module.scss
@@ -2,6 +2,7 @@
   display: flex;
   flex-direction: column;
   text-align: justify;
+  padding: 0 2rem;
 
   h1 {
     font-size: 2rem;
@@ -34,8 +35,9 @@
   }
 
   .meta {
-    text-align: right;
-    margin: 0.1rem initial;
+    padding-bottom: 4rem;
+    text-align: left;
+    font-size: 1.2rem;
     color: #888;
   }
