+++ styles/global.css
@@ -1,5 +1,5 @@
 :root, [data-color-mode=light] {
-  --color-template-background: #e5e5e5;
+  --color-template-background: #ffffff;
   --color-anchor: #0070f3;
   --color-visited-anchor: #027148;
   --color-table-border: #333333;
