+++ styles/home.module.css
@@ -2,7 +2,7 @@
   display: flex;
   flex-direction: column;
   width: 100%;
-  padding: 0 1.5rem 2rem 1.5rem;
+  padding: 0 2rem;
   font-size: 1.2rem;
   line-height: 1.5;
 }
